move cbm_bits checks to resctrl
