+++ owca/containers.py
@@ -31,7 +31,6 @@ from owca.metrics import Measurements, MetricName
 
 log = logging.getLogger(__name__)
 
-log = logging.getLogger(__name__)
 DEFAULT_EVENTS = (MetricName.INSTRUCTIONS, MetricName.CYCLES, MetricName.CACHE_MISSES)
